`_stack_onto_` fix

+++ torchrl/data/tensordict/memmap.py
@@ -8,12 +8,12 @@ from __future__ import annotations
 import functools
 import os
 import tempfile
-from math import prod
 from typing import Any, Callable, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
 
+from torchrl import prod
 from torchrl.data.tensordict.utils import _getitem_batch_size
 from torchrl.data.utils import (
     DEVICE_TYPING,
@@ -253,12 +253,15 @@ class MemmapTensor(object):
         self,
         idx: Optional[int] = None,
         memmap_array: Optional[np.ndarray] = None,
+        from_numpy: bool = False,
     ) -> torch.Tensor:
         if memmap_array is None:
             memmap_array = self.memmap_array
         if idx is not None:
+            if isinstance(idx, torch.Tensor):
+                idx = idx.cpu()
             memmap_array = memmap_array[idx]
-        out = self._np_to_tensor(memmap_array)
+        out = self._np_to_tensor(memmap_array, from_numpy=from_numpy)
         if (
             idx is not None
             and not isinstance(idx, (int, np.integer))
@@ -269,7 +272,9 @@ class MemmapTensor(object):
             out = out.view(size)
         return out
 
-    def _np_to_tensor(self, memmap_array: np.ndarray) -> torch.Tensor:
+    def _np_to_tensor(self, memmap_array: np.ndarray, from_numpy: bool) -> torch.Tensor:
+        if from_numpy:
+            return torch.from_numpy(memmap_array)
         return torch.as_tensor(memmap_array, device=self.device)
 
     @classmethod
@@ -292,6 +297,11 @@ class MemmapTensor(object):
     @property
     def _tensor(self) -> torch.Tensor:
         return self._load_item()
+
+    @property
+    def _tensor_from_numpy(self) -> torch.Tensor:
+        # a tensor created with `from_numpy` to make sure that changes are done in-place
+        return self._load_item(from_numpy=True)
 
     def ndimension(self) -> int:
         return self._ndim
@@ -382,10 +392,8 @@ class MemmapTensor(object):
         return self
 
     def __del__(self) -> None:
-        # if hasattr(self, "filename"):
-        if self._has_ownership:
+        if "_has_ownership" in self.__dir__() and self._has_ownership:
             os.unlink(self.filename)
-            # self.file.close()
 
     def __eq__(self, other: Any) -> torch.Tensor:
         if not isinstance(other, (MemmapTensor, torch.Tensor, float, int, np.ndarray)):
@@ -398,7 +406,9 @@ class MemmapTensor(object):
                 attr
             )  # make sure that appropriate exceptions are raised
 
-        if attr not in self.__getattribute__("_tensor_dir"):
+        if ("_tensor_dir" not in self.__dir__()) or (
+            attr not in self.__getattribute__("_tensor_dir")
+        ):
             raise AttributeError(f"{attr} not found")
         _tensor = self.__getattribute__("_tensor")
         return getattr(_tensor, attr)
@@ -450,8 +460,10 @@ class MemmapTensor(object):
         return self._load_item(idx=item)
 
     def __setitem__(self, idx: INDEX_TYPING, value: torch.Tensor):
-        # self.memmap_array[idx] = to_numpy(value)
-        self._load_item()[idx] = value
+        if self.device == torch.device("cpu"):
+            self._load_item()[idx] = value
+        else:
+            self.memmap_array[idx] = to_numpy(value)
 
     def __setstate__(self, state: dict) -> None:
         if state["file"] is None:
@@ -538,7 +550,11 @@ def stack(
     list_of_tensors = [
         a._tensor if isinstance(a, MemmapTensor) else a for a in list_of_memmap
     ]
-    return torch.stack(list_of_tensors, dim, out=out)
+    if isinstance(out, MemmapTensor):
+        list_of_tensors = [tensor.cpu() for tensor in list_of_tensors]
+        return torch.stack(list_of_tensors, dim, out=out._tensor_from_numpy)
+    else:
+        return torch.stack(list_of_tensors, dim, out=out)
 
 
 @implements_for_memmap(torch.unbind)
